🐛 fix(trafficLight.asl): add newBid rule to handle already existing bids ✨ feat(Env.java): change weight parameter type to long to handle larger values 🚀 chore(MainFrame.java, Simulator.java, TrafficLight.java, Vehicle.java): change carAvgWaitingByDirection type to Double to handle decimal values and fix related methods

+++ junctionManagement/src/env/main/Env.java
@@ -129,7 +129,7 @@ public class Env extends Environment {
      * Notifies the agent that a new vehicle has arrived
      * @param direction The direction of the traffic light
      */
-    public void notifyAgentNewBid(Direction direction, int weight, int count){
+    public void notifyAgentNewBid(Direction direction, long weight, int count){
         addPercept(agentNames.get(direction),Literal.parseLiteral("newBid").addTerms(new NumberTermImpl(weight), new NumberTermImpl(count)));
         informAgsEnvironmentChanged(agentNames.get(direction));
     }
